Update Topmenu
Updated topmenu to match website text. Moved logout to dropdown menu. Added account link for changing account information. Not active as there is not a page for it.

Updated lookupReservation.css to better match rest of website.

Added check on reservation page for if a user logged in. If there is no user logged in, the user can still look at prices, but will not be able to submit the reservation.

Added same check to Registration.jsp

Updated .war file